third_party/WebKit/LayoutTests/imported/wpt/html/infrastructure/common-dom-interfaces/collections/htmlformcontrolscollection-expected.txt

This is a testharness.js-based test.
FAIL The length attribute must return the number of elements in the form assert_equals: The length attribute is incorrect. expected 2 but got 1
PASS HTMLFormControlsCollection.item(index) must return the indexed item 
PASS HTMLFormControlsCollection[index] must return the indexed item 
PASS HTMLFormControlsCollection is not callable 
PASS HTMLFormControlsCollection[name] must return the named item 
PASS HTMLFormControlsCollection.namedItem(name) must return the named item 
PASS The namedItem(name) must return an Element 
PASS The namedItem(name) must return RadioNodeList 
PASS The namedItem(name) must return null if the name is empty 
PASS The namedItem(name) must return null if there is no matched element 
FAIL Controls can be indexed by id or name attribute assert_equals: Controls can be named by 'id' attribute. expected Element node <keygen id="kg" name="key"></keygen> but got null
PASS The namedItem(name) must return the items with id or name attribute 
FAIL The HTMLFormControlsCollection interface is used for collections of listed elements in form element assert_equals: The length should be 8. expected 8 but got 7
FAIL The HTMLFormControlsCollection interface is used for collections of listed elements in fieldset element assert_equals: The length should be 8. expected 8 but got 7
PASS The controls in the form element must be sorted in tree order 
Harness: the test ran to completion.
